requirements.txt
# Core
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tenacity>=8.0.0
instructor>=1.0.0
openai>=1.0.0

# Scrapers & Tools
yt-dlp>=2023.0.0
scrapetube>=2.0.0
feedparser>=6.0.0
DrissionPage>=4.0.0
twikit>=1.5.0
bilibili-api-python>=17.0.0
